Add numbers to the detailed steps section
It was not clear to me that the numbered list was simply a summary of the steps that followed. It looked like you did the first 6 things and then you did the following 5 things.

I've numbered the detail sections to try and connect them to the summary list so they hopefully look more connected. I would probably remove the summary altogether since the titles on the detail section contains the same information and it's not a process that you would need an index for.