More information about SendGrid
Because SendGrid does not automatically approve everyone for use of their services there could be a delay in setting up email. This could be a real blocker for a newly launched site if email is critical, such as a member registration site. I did not know this and my launch is now being held up.