Include www redirection service content and cname flattening within redirect-incoming-requests.md
You can't use the second snippet to standardize on a bare domain without a CNAME flattening service.  So, we need to add the steps to set that up

Adding link to Pantheon Cloudflare doc

If a person has an SSL cert, they don't need the Cloudflare setup.  Adding that disclaimer.