Update documentation on sql-sync and sql-sync-pipe
sql-sync-pipe has a bad habit of producing empty databases on Pantheon. sql-sync, on the other hand, is working well as of Drush 8.0.4 (not yet available). We could hold this documentation update until 8.0.4 is out, or just push it out and let folks figure it out. Note that an alternative solution is suggested.