More Info on Site Freezing
I'm iffy on this part: "Should there be any issues with the un-freezing process, a backup of the site is available to be restored via the site dashboard." 
But we are typically seeing a few cases of it per day. We can handle it on our end, but it would be good if there were something here too.