Add Drupal 8 setting config snippet
The Drupal configuration appears to be just for D7 sites. I had to tinker around/read the D8 module's README.md in order to get the module to recognize the library and become enabled. I am not sure this is the best/recommended method but it appears to work, at least for me.