Update 01-introduction.md
Many experienced WP devs use FTP, so we don't want to imply they are beginners. This guide is targeted at the ones who are happy not using git, but don't realize how easy it is to use FTP on Pantheon. Kind of a "barely use git" or "use git without even knowing it" approach. LMK if that makes sense.