Update logs.md
Expand "60 days of logs" (line 29) to the more precise "Up to 60 days of logs".

A customer asked how they could access all 60 days of their nginx-access.log based on the information shown here, which is misleading / inaccurate information as we only keep the logs until the appserver they're on is migrated. If their appserver has just been migrated, they may only have a few hours (or less) of logs available.