Update cookies.md
The only time this crucial requirement is mentioned on this page it sounds as if it's talking about a specific use case for cookies related to serving cached content. I propose making it very prominent at the top that this is required, since any use of cookies seems to depend upon that naming scheme.