Git Rebase -> Git Merge for Custom Upstream Updates
This is a change in how we advice folks to update custom upstream code. Previously we used rebase but it could cause conflicts. Merge is a lot cleaner. Reference: https://www.atlassian.com/git/tutorials/merging-vs-rebasing

I updated the doc to use merge instead and believe the syntax should be the same as per https://git-scm.com/docs/git-merge.