Remove erroneous wording for team management
Suggesting to remove "and select the applicable plan for the site" from the end of the description in the second sentence under Paid Sites, as that is not the screen encountered by myself or a client when I follow these instructions. The option to select the applicable plan for the site is absent.