Update redis.md
Redis as of 3.13 supports compressing the cache using the `Redis_CacheCompressed` class.  This saves a whopping 80% of memory on Redis for a small tradeoff on CPU useage.  For large sites this is huge, especially when trying to fight high Redis turnover rates.  I implemented this on one of our large sites on Pantheon and noticed an actual 75% memory reduction.  I hope to increase awareness of this great feature.