Removing depreciated log files from list
`watcher.log` and `pyinotify.log` have been depreciated, and should not be listed on this doc, as they are no longer accessible to the customer. `mysqld.log` and `mysqld-slow-query.log` are found in the Database `logs/` directory, not the app container, so I added language to specify this. Verified by Josh Koenig in Product.