Clarify that the Platform API on Drupal 8
The existing text caused some confusion for a user who thought that no API module being available for Drupal 8 meant that there was some missing functionality for this version of Drupal. Clarified that the API module is not available because it is not necessary, and that this module is not intended to be used as an API by end users (and therefore probably should not have been named "API").