Removing IC Prefix Requirement for D9 + IC
Since we have the organizational feature flag for Drupal 9 + Integrated Composer, we don't need folks to actually prefix their site with ic-demo-2020 anymore. This change just removes that requirement but (a) still keeps the NOT FOR PRODUCTION USE vibe and (b) will still require them to spin it up in the "Drupal 9 Early Access" group which prevents the site from going live.