Update plugins-known-issues.md
Adding more information to All-in-one WP Migration plugin. This plugin causes serious headaches, most of the time the site has to be migrated to a new sandbox or platform engineering has to get involved. This causes all deployments from Dev to be blocked, and usually takes over 24 hours to resolve.