Finishing out the "Deploy" process.
Customers have come into support noting that the current documentation does not completely get a working multidev and need further instructions to get it working. This is my first attempt to finish this documentation. A few customers have gone through these steps and have verified that it did work. But I am also certainly open to feedback or other ways to improve this for all users.