Show that Search API Pantheon has Drupal 9 support
The docs are outdated because the [Search API Pantheon](https://www.drupal.org/project/search_api_pantheon) 3.0.0 release has a Drupal 9 support. Possibly something should be mentioned here about the need to update Search API Solr to the ^4.2 and install the legacy sub module if on Solr 3.x on Pantheon.