Update bash install script
The code that selects the version to install terminus from assumes that the most current version i.e 3.x versus 2.x is listed first.   If you are assuming the 3.x code is what is going to be downloaded and you have further 3.x specific features like plugins used then your script will fail.  

This designates a known fallback version if for some reason the 2.x release is listed first.