Update errors-and-server-responses.md
Requesting to add:
--
### Error 404 Not Found

Most times it will show `Failed to load resource: the server responded with a status of 404 ()` if the element is inspected from a web browser and a lot of different variations.

Basically, it means that the client (via web browser) was able to successfully connect to the host (website’s application server), but it was unable to find the actual resource that was requested (i.e. a specific URL or filename). So the best way is to confirm if the URL and file do exist.

--

This is to prevent confusion coming from the other 404 
And made me initiate a request to update the doc due to this ticket: https://pantheon.zendesk.com/agent/tickets/679199 

Any relevant amendments are welcome. Thank you team.