Rephrasing of sentence for better clarity
Hey friends,
As part of my Pantheon CSE onboarding I need to submit one change to a public facing document via Git Pull Request, so I scrolled through some of the documents and found a sentence that could be cleaned up slightly (it leverages "as it is" to reference the first clause, and "for which", both of which add unnecessary sentence complexity). 

Feel free to reject this if it doesn't fit with the preferred writing tone, but as part of my onboarding I needed to contribute back to the public docs. Have a lovely weekend!