Fixing incorrect link within body copy
Was reading through the Solr docs and following the links throughout the page when I found that one of the links took me to the incorrect location. This patch should address the issue and link to the correct location, which is slightly further down the page. 

Context: “[Check Index and Batch](link) sizes for possible scenarios or more time-outs.”
- Incorrect link: https://pantheon.io/docs/solr#apache-solr-vocabulary
- Correct link: https://pantheon.io/docs/solr#check-index-and-batch-sizes