Update plugins-known-issues.md (#8550)
* Update plugins-known-issues.md

I’m seeing that this part of the wordfence issue of our doc might be misleading when the customers are creating their symlinks. Since the symlink commands in the doc are assuming that our customers are in their /code directory and not in /code/wp-content directory.

In step 3, the code snippet for that is correct. However, stating /wp-content is a bit misleading.

* copy edits and formatting

* updated code example

---------

Co-authored-by: ravenloveriel-pantheon <103234849+ravenloveriel-pantheon@users.noreply.github.com>
Co-authored-by: whitneymeredith <97134315+whitneymeredith@users.noreply.github.com>