Update 02-agcdn-features.md
Removing the reference to URL Rate Limiting on this general feature page for AGCDN. Some customers found it confusing or misleading to reference that feature on a page seemingly focused on the standard suite, not full suite. Also, the details of that capability are already on the WAF+IO full suite guide page.