Update cli.md (#8787)
This patch addresses two issues. The first is an incorrect URL to a headline on the same page that was not updated to reflect a word change (breaking the link). The second is removing an extra comma where it grammatically isn't needed.

Link fix:
Old: `/object-cache/cli#clearing-cached-data`
New: `/object-cache/cli#clear-cached-data`