Add
Pantheon outlines that the Object Cache (and Object Cache Pro) plugins are only available for certain plans. We then go on to explain that there are alternatives available, but we do not recommend a specific option.

This update adds a Pantheon-owned plugin as a recommended alternative, while making it clear that other plugins can still be used instead.