Added note about adding subdomains of pantheonsite.io 03-platform-domains.md
Edge Routing recently put a limitation in place that prevents customers from adding custom domains in the pantheonsite.io namespace. https://getpantheon.atlassian.net/browse/EDRT-4140

While it was never recommended or prescribed, there are a handful of customers that were engaging in this practice who are now confused about the limitation.