Add clarity to WPMS documentation about requesting multisites (#9332)
* add explanation about why WP sites can't just magically become multisites

* update the custom upstream intro to add glossary terms
I am hypothesizing that glossary terms are added based on the <dfn> tag. I am further hypothesizing that if we leave the markdown syntax on this page intact, the new glossary terms would not be added correctly, therefore I have converted all the markdown bullets to html list items.

I have added "framework" on this page because it makes the most sense to be included on a page about upstreams.

* a little more clarity about what settings

* remove the note, rephrase

* remove html formatting
didn't work for the glossary

* use headings and p tags

* use upstream framework as the term instead

* rename upstream framework to site framework
more in line with what exists in the dashboard

* site framework doesn't create the glossary page
go back to upstream framework but describe as "site framework" in parenthetical

* use em tags for emphasis

* link to framework in the glossary

* add space

Co-authored-by: Phil Tyler <phil.tyler@pantheon.io>

* fix the link

* clarify wpms upstream creation

Co-authored-by: Rachel <rachel@pantheon.io>

* simplify existing wpms conversion description

Co-authored-by: Rachel <rachel@pantheon.io>

* support creates the upstream, not the multisite itself

* Minor phrasing change to WPMS upstream access

* don't say "wordpress network"

* Update source/content/guides/multisite/01-introduction.md

Co-authored-by: Rachel <rachel@pantheon.io>

* Apply suggestions from code review

Co-authored-by: Rachel <rachel@pantheon.io>

---------

Co-authored-by: Phil Tyler <phil.tyler@pantheon.io>
Co-authored-by: Rachel <rachel@pantheon.io>
Co-authored-by: Chris Charlton <chris.charlton@pantheon.io>