Update modules-known-issues.md (#8836)
* Update modules-known-issues.md

There are no more known issues for the module:
"Simple OAuth / OAuth 2.0
Issue: The module requires a very specific set of permissions for the folder and the keys to be uploaded. Using Private or non-standard filepaths won't work. It is not possible to change these in LIVE or TEST environment.

Solution: You can try to patch the permission check in the module. The alternative is to use off-site key management tools like Lockr"

The module now works without a workaround.

* Revert "Update modules-known-issues.md"

This reverts commit c539023cfd4ee0752dff9394887cd754c566198a.

* Add known solution for OAuth permissions issue

* Update source/content/modules-known-issues.md

Co-authored-by: Steve Persch <stevector@users.noreply.github.com>

* Update source/content/modules-known-issues.md

Co-authored-by: Steve Persch <stevector@users.noreply.github.com>

---------

Co-authored-by: Rachel Whitton <rachel@pantheon.io>
Co-authored-by: Steve Persch <stevector@users.noreply.github.com>