Updating to include Site traffic per Billing Cycle
New feature added to the site list page, that allows you to view site traffic per billing period. 

Worry that this feature is going to make customer believe that we are coming back with overages, so included a call out that we are releasing this not for overages but so customers can better provision their sites.