clarify instructions for adding upstream dependencies (#9576)
Tweaked the wording slightly to make it clear that you're switching to a directory, and affecting a file via commands. 

Removed mention of setting a "version" as it's very confusing and after asking internally, Pantheon devs confirmed: "After a conversation with the team about this I would say the docs are at best confusing, and at worst not entirely accurate on this point. In practice you can just pin packages to specific versions in the usual way, if necessary." -- mentioning this version refers to composer package version.